R101-003 Revisit definition of effectively volatile types and objects
There is no reason to treat types and objects with Async_Writers and
Effective_Reads properties set to False as effectively volatile. Make this
distinction now so that only objects whose evaluation should be considered
as having an effect are restricted in their usage.

This requires separating the notion of "effectively volatile" from the
narrower notion of "effectively volatile for reading", and use that new
notion for those types and objects where a read should be considered
as having an effect.

Change-Id: I2ffeab30cd5ebbcc35f8509a759a93150fb04982