SA25-033 avoid duplicated checks on calls with invariant checks
To emit an invariant check, we reuse the parameters of the call. This
may cause RTE checks for the parameters to be duplicated.

* src/why/gnat2why-expr.adb
(Compute_Call_Args): New parameter Use_Tmps to force the addition of
 temporaries for parameters
(Transform_Function_Call): Set Use_Tmps to True if we need to check
 invariants on call. Do not assume that no temporaries were needed for
 the call.
(Transform_Statement_Or_Declaration): Set Use_Tmps to True if we need to
 check invariants on a procedure or entry call.

Change-Id: I85e904f6bb302ef9d3f05ecfa5c07fd02d2bc933