SA13-005 Add reasons for checks
Index/overflow/length/range checks may be located where a user finds it
odd to have such a check, due to the combination of SPARK rules and
location on the token for the top-level sub-expression (like symbol +
for an addition). Thus, we add here an explanation in the form of a
"reason for check" part of some of the check messages, similar to other
explanations whose role is to help in understanding the cause for a
message.

* src/flow/flow_error_messages.adb
(Get_Details): New function to get a reason for some of the check messages.
(Error_Msg_Proof): Add a reason for some of the check messages.

* testsuite/gnatprove/tests/SA13-005__reason_for_check/reason_for_check.adb
* testsuite/gnatprove/tests/SA13-005__reason_for_check/test.out
* testsuite/gnatprove/tests/SA13-005__reason_for_check/test.py
Add test.

* testsuite/gnatprove/tests/T623-013__output_checks/test.out
Update output. Only runs with --output=oneline have the new information.

* testsuite/gnatprove/tests
Update output for many tests, as --output=oneline is the default.

Change-Id: I7193a15a83f5ae63d07257e72c27f8ef0903d4df