T527-057 add test about aliasing of previously synchronized pointers
This test was crashing proof, because aliasing check was not applied to a
constant pointer (because all constants were considered to be synchronized).
Now aliasing is detected in flow and proof appears to work as expected.

* testsuite/gnatprove/lib/python/test_support.py
(get_tag): Reorder if-elif chain to avoid spurious detection in message
<<formal parameter "X" and global "My_Global" are aliased>>, which matches
both GLOBAL and ALIASING tags.

* testsuite/gnatprove/tests/T527-057__flow_aliased_access_constant/
Add; aliasing is detected.

Change-Id: Iabfe73806c965e1a353450ced5e6ad2ecee31182