T623-010 Rephrase hints at a possible fix in messages
Explanations are not the same as hints. The latter are typically phrased as
something that should be done to fix the problem. Introduce these by
"possible fix" instead of "possible explanation", which we keep for actual
explanations.

* src/flow/flow_error_messages.adb
(Get_Fix): Renaming of Get_Explanation. Skip for static checks on
 Unchecked_Conversion.
(Error_Msg_Proof): Separately issue an explanation and a fix when present.

* testsuite/gnatprove/tests
Output update.

Change-Id: Ife140707ad34ca90419e88b73aec072deaeb5349