T608-010 Subprograms whose address is taken are not always inlined
Contextual analysis in GNATprove relies on the fact that it is known which
subprograms are always inlined. This can't be the case for subprograms whose
address is taken, which should be taken into account in GNATprove. Note that
this does not prevent direct calls to such subprograms from being inlined.

* src/spark/spark_util-subprograms.adb
(Is_Local_Subprogram_Always_Inlined): Do not consider a subprogram as always
 inlined if its address is taken.

* testsuite/gnatprove/tests/T608-010__callback_inlining
Add test.

Change-Id: I249381fa07b9a13b9e72af20a5da67f651fb386c