T618-021, R423-011 deconstruct hidden --flow-termination switch
With improved diagnostic for nonterminating subprograms and --proof-warnings
switch there is no longer need for a hidden --flow-termination switch.

* src/common/gnat2why_opts.ads
* src/flow/flow-analysis.adb
* src/flow/flow-analysis.ads
* src/flow/flow.adb
* src/gnatprove/configuration.adb
* src/gnatprove/gnat2why_opts-writing.adb
* src/spark/gnat2why_opts-reading.adb
* src/spark/gnat2why_opts-reading.ads
Remove code related to --flow-termination switch.

* testsuite/gnatprove/tests/P715-019__flow_generic_termination/
* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination/
* testsuite/gnatprove/tests/P929-043__flow_pkg_elaboration_in_direct_calls/
* testsuite/gnatprove/tests/PB25-025__flow_termination/
* testsuite/gnatprove/tests/Q210-039__flow_calls_in_assignment/
* testsuite/gnatprove/tests/Q512-013__flow_type_aspects/
Replace --flow-termination switch in test.py files with Terminate annotation
and update outputs. Many diffs, but all of them seems legitimate.

Note: --flow-termination causes termination to be checked *without* callers
assuming that the subprogram is terminating; Terminate annotation is forcing
checks *with* callers assuming this annotation to be correct (except for
mutually recursive callers and even more exceptionally when
inlining-for-proof kicks in).

Change-Id: I61272d48184604a4b9814f308b1857d9bd96a327