SA25-033 add test for variants on recursive data-structures
I tried to come up with a way to prove the termination on length on
lists. It can be done by adding an extra parameter, but this parameter
is harmful for proof. It is possible to restate the usual definition
through a lemma, but then the lemma needs to be manually instanciated.
The result of this tentative, is that we probably need a structural
variant if we want to be able to prove termination of Length without
decreasing provability.

Change-Id: I93467346c3411e20ed8554ca11af21a1bb319bf8