T727-032 Add possible fix message part for using expression functions
When a function without contract (explicit or implicit) is called in an
expression submitted to a check (whether it is a run-time check or an
assertion), and the check cannot be proved, it is very likely that a
postcondition needs to be added to the function, or the function needs to
be turned into an expression function. Add a possible fix message for this
case, that works for direct calls, dispatching calls and indirect calls.

* src/flow/flow_error_messages.adb
(Get_Fix): Add fix message for this case, when no prior fix was proposed
 based in particular on variable occurrences in the expression submitted to
 a check.

* src/flow/flow_types.adb
(Flow_Id_To_String): Move local function Get_Operator_Symbol to other unit.

* src/spark/spark_util.adb
* src/spark/spark_util.ads
(Get_Operator_Symbol): New query moved from flow_types.adb.
(Source_Name): Improve the name for operators, to avoid using the mangled
 name internally created.

* testsuite/gnatprove/tests/T727-032__fix_expr_fun/exprfun.adb
* testsuite/gnatprove/tests/T727-032__fix_expr_fun/exprfun.ads
* testsuite/gnatprove/tests/T727-032__fix_expr_fun/test.out
* testsuite/gnatprove/tests/T727-032__fix_expr_fun/test.py
Add test.

* testsuite/gnatprove/tests
Update test outputs.

Change-Id: I7540d79de0ac64f73a838a6d71c2cc6a608389fd