T716-054 Make it possible to open all tests in one project
It is quite convenient during trainings to be able to open all snippets
of code used in the SPARK UG to reproduce analysis locally. This is now
possible again with an aggregate project under the tests directory.

* docs/ug/examples/tests/bank2/account2.ads
* docs/ug/examples/tests/bank3/account3.ads
* docs/ug/examples/tests/bank4/account4.ads
* docs/ug/examples/tests/bank5/account5.ads
* docs/ug/examples/tests/bank6/account6.ads
* docs/ug/examples/tests/bank7/account7.ads
* docs/ug/examples/tests/bank8/account8.ads
Add missing files.

* docs/ug/examples/tests/gnatprove_by_example.gpr
Aggregate project for all snippets in SPARK UG. One can open a file and
use GNATprove on that file. Note that messages are not properly located when
a file name is present in multiple projects.

* docs/ug/examples/tests/test.adc
As Builder packages of subprojects are ignored in an aggregate project,
we need to repeat here the default pragma configuration file.

* docs/ug/examples/tests/main3/lights.ads
* docs/ug/examples/tests/main3/main3.adb
* docs/ug/examples/tests/main3/test.gpr
* docs/ug/examples/tests/main3/test.out
* docs/ug/examples/tests/main3/test.adc
Useless test removed.

Change-Id: Ia4e15d3270902d518c0e2850bd4f6e5a10dfd335