T725-001 no empty output in testsuite
If an empty output file is included in the doc, doc generation
crashes. That's why the automatic update script aborts when it
encounters empty new baselines. So we change the testsuite so that
empty baselines are represented by the string "(no output)".

As a technical limitation of the current scripts, an empty baseline
is *also* accepted as correct baseline, but this may change.

* run_examples.py
  Output refiner that creates the output "(no output)" in case of
  empty output

Change-Id: I38bf4a9fd02aa4d6ee8251c9f0e45e97d93514c9