T610-004 Add helpful functions for float from/to integer conversion
Augment the SPARK lemma library unit on floats with helpful functions
used for proving properties of the conversion between floats and integers.

As a workaround, -gnateDSPARK_BODY_MODE=Off is passed explicitly in two
tests, which should not be needed as the instances used are part of the
lemma library for which the project file specifies this compilation switch
already. To be removed once the underlying issue is fixed.

* include/spark-floating_point_arithmetic_lemmas.ads
Add new useful ghost functions, taking into account the precision of the
floating-point type being instantiated.

* testsuite/gnatprove/tests/T610-004__float_arithmetic_lemmas
Add test to exercise float lemmas.

* testsuite/gnatprove/tests
Update tests. Increase steps to prove checks from the lemma library.

Change-Id: Ic473d0dc57939d75115132e796d6c34cd5dc0fe4