minor replace box with explicit default value
When applying GNATstudio menu Build->Check Semantics (which is equivalent to
-gnatc) on unit gnat2why-expr.adb, GNAT is emitting a spurious warning
saying that box in Binder_Type aggregate is not initialized.

This is annoying, though the underlying GNAT problem is quite likely not
worth the effort, so just rewrite the code with explicit default value.

* src/why/gnat2why-expr.adb
(Generate_Logic_Function): Replace <> with explicit default value.

no-tn-check

Change-Id: Ie6d1248531b800be004ac24ad36390e11ebf0994