T804-037 pull routine given by Iterable_For_Proof for translation
If the Iterable_For_Proof annotation appears in a WITH-ed unit, the function
it specifies must be pulled for translation in marking, as it will be later
needed for proof.

* src/spark/spark_definition-annotate.adb
(Check_Iterable_Annotation): Pull function given by the Iterable_For_Proof
for translation.

* testsuite/gnatprove/tests/T804-037__functional_map/
Crash is gone, so enable and add output.

Change-Id: Ib2530d3409e699561b0f15349b17a1251e72d519