T812-027 avoid unnecessary processing of calls for DDG
When generating globals, we represent calls to subprograms with potential
outputs (which means procedures and entries with no usable Global/Depends
contract) as variables. This way DDG tells us whether those calls happen
definitely or conditionally (just like it tells us which variables are
always written and which only conditionally).

This processing was done for all subprogram calls, including calls to
functions (which cannot have outputs), calls to procedures with usable
Global/Depends (whose effects are precisely represented in flow graphs)
and even when not generating Global contracts.

Now we only do this processing when required. Behaviour is not affected,
though performance should now be better.

* src/flow/flow-data_dependence_graph.adb
(Create): Only process subprogram calls when necessary.

Change-Id: Ia85702b78d0297cd30a60ea2b4e5e7b1590ebae5