T812-027 fix dependency tracking for record self-assignments
Data flow within a record self-assignment like:

   Tmp := (A => Tmp.B, B => Tmp.A);

was represented as a series of assignments to record components, i.e.:

   Tmp.A := Tmp.B;
   Tmp.B := Tmp.A;

but it is not possible to represent a self-assignment as a sequence of
assignments with no temporary variables. Now this self-assignment is
represented as a sequence of reads (VU, or variables used) followed by a
sequence of writes (VD, or variables used) with edges added according to
data flow relation:

   VU : {Tmp.B}--+
   VU : {Tmp.A}--|--+
   VD : {Tmp.A}<-+  |
   VD : {Tmp.B}<----+

This is inspired by how 'In and 'Out views of procedure call parameters
are represented with dedicated vertices in CFG and connected with edges
according to the Depends contract in TDG.

* src/flow/flow-control_flow_graph.adb
(Do_Assignment_Statement): Add dedicated vertices for reads and writes
of a record component.
(Do_Call_Statement): Tune comment with a similar used of Splice like above.

* src/flow/flow-data_dependence_graph.adb
(Create): Add edges for record assignments.

* src/flow/flow.adb
(NDI): Adapt flow graph pretty-printing routine to record assignments.

* src/flow/flow.ads
* src/flow/flow_types.ads
(Flow_Graphs): Move to Flow_Types, so its Vertex_Id type can be used as
a flow graph vertex attribute.

* testsuite/gnatprove/tests/T812-027__flow_record/test.out
Spurious checks are gone.

Change-Id: I3c2c9256bbaf962cccd70966785ad11842330629