T820-029 remove redundant code for string literals
The code which is removed here inserts the constant values for
'first, 'last and 'length of string literals. But the theories for
string literal subtypes already have constants for these values, so
the general code which refers to those constants also works. Testing
shows that referring to the constant rather than the concrete value
makes no difference for proof.

* why-gen-expr.adb
(New_Attribute_Expr): special handling of string literal subtypes
  removed

Change-Id: I1b39d388daf8fd01fd3528baeda461eb84719e8b