T827-002 clarify result of Compute_Ada_Node_Set with a contract
Routine Compute_Ada_Node_Set returns entities (which are translated to
individual Why3 modules because this is elegant), aggregates (which are
complex enough to be translated into individual Why3 modules as well) and
Access attributes. This is now explicitly stated in the postcondition.

Also, refine the guard in a recent fix where this routine was returning an
aggregate, because apparently it can also return other non-entity nodes.

* src/why/why-inter.adb
(Compute_Ada_Node_Set): Add postcondition.

* src/why/gnat2why-expr.adb
(Assume_Dynamic_Invariant_For_Variables): Refine guard.

Change-Id: I32068fa1adb15e5fbb5a504b15eca87484c5020e