T910-001 components of init wrappers for records might not be wrappers
It is possible for a component of an init wrapper type for records to
have components which are not wrappers. This occurs when the record type
is not used by any object which are annotated with relaxed initialization
but only for intermediate expressions.

* src/why/gnat2why-expr.adb
(Transform_Record_Component_Associations): Do not use wrappers if the
 component type cannot have relaxed initialization.

* src/why/why-gen-records.adb
(W_Type_Of_Component): Do not use wrappers if the component type cannot
 have relaxed initialization.

* tests/T910-001__relaxed_init
Enable test.

Change-Id: Ifc774c61146d201edecc099ad5a5bac14cd31877