T904-001 Update test after changes in CVC4
With the new release of CVC4, one fewer check was proved with 5000 steps.
Use that opportunity to revisit choices for this test, and reenable Z3 and
Alt-Ergo prover, leading to two more checks proved instead.

Also reenable counter-examples, which should all be spurious on this test
and a good target for filtering in the future.

Change-Id: I414827976901f0e0021f1f2f212241e8382dd4fd