T921-021 do not assume that contract cases are always boolean
When we need an initialization wrapper for standard boolean, we do
not translate it using the Why3 boolean type, so that we can have
types of kind EW_Abstract and EW_Split that can store a flag for
relaxed initialization. Though expressions of contract cases can never
have relaxed initialization, they will be translated to a named type
in that case. Therefore we should not assume that their type is
necessarily EW_Bool_Type.

* src/why/gnat2why-expr.adb
(Transform_Attribute_Old): An old expression in the predicate domain
 might not have EW_Bool_Type as an expected type.
(Compute_Contract_Cases_Exit_Checks): The consequence of a contract case
 might not have EW_Bool_Type as an expected type.

Change-Id: I7dac342282edadf470a87e31aaf5c8dcffc1caa3