T921-020 relax checks on source of unchecked conversion
We differentiate between the checks of a source and the target of an
unchecked conversion, and relax the checks on the source. Invalid
bitpatterns are not an issue anymore, but holes still are, though
they are only relevant for composite types.

We remove the previous VC Kind VC_UC_No_Holes and replace it by
VC_UC_Source and VC_UC_Target. VC_UC_Target is also used for
overlays, as it is the stronger of the two and overlays may be both
source or target.

* vc_kinds.adb
  removed old VC Kind, added two new ones
* flow_error_messages.adb
  (Get_Severity): adapt to new VC Kinds
* spark_report.adb
  (VC_Kind_To_Summary): adapt to new VC Kinds
* spark_util-types.adb
  (Check_Composite_Type_No_Hole): generic helper for composite types
  (Type_Has_No_Holes): new function to compute if type has a hole
* gnat2why-error_messages.adb
  (Make_Codepeer_Msg): adapt to new VC Kinds
  (Not_Proved_Message): adapt to new VC Kinds
  (Proved_Message): adapt to new VC Kinds
* gnat2why-expr.adb
  (Transform_Declaration): use VC_UC_Target instead of
  VC_UC_No_Holes
* gnat2why-subprograms.adb
  (Generate_VCs_For_Subprogram): use VC_UC_Source for source and
  VC_UC_Target for target of unchecked conversion

Change-Id: I8da1139ed01b48be21ab454a6b42f60bfb7248c8