S701-026 Add index to SPARK User's Guide
Add an index to the SPARK User's Guide. When multiple entries share the same
key in the index, qualify all but one (at most) with a context as in:

.. index:: precondition; example of use

Also use "pair" for mutual contexts between two keys, and "see" for
indirection between keys.

For details of the syntax and possibilities, see:
https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html

* rst files
Add index entries.

* examples/tests/addition_bignum
New example to remove use of -gnato13 in this basic example.

* examples/tests/addition_rewrite
* examples/tests/addition_saturated
Rename main file to "addition.adb" now that all tests are in their own
subdir.

Change-Id: I2deeb13be38d807f35edcac0ee45427673c46ec3