QB27-015 extend crash message to all kinds of calls resolved in GG
It was a mistake to only crash on definite calls to abstract subprograms
when GG gives up (because it cannot know what the actual callee is and there
are no tasking-related contracts on abstract subprograms). Now it also
crashes on conditional and proof calls.

This fixes an utterly confusing crash on M402-001__mvm, where a conditional
call to an abstract subprogram happens in an obfuscated source code with a
complex hierarchy of generic units.

* src/flow/flow_generated_globals-partial.adb
(Categorize_Calls): Guard all accesses to "Contracts (Pick)" with check for
abstract subprograms.

Change-Id: I2a2dd8e55ffc341c9f8be2f443802e5043a7b15b