T826-013 add test with child packages replaced with nested ones
Add test with the same structure of abstract state hierarchy, but attached
to nested packages, so the GG algorithm solves everything in phase 1 for
a single compilation unit. The original tests involves 4 compilation units,
so the GG algorithm resolves inter-unit subprogram calls in phase 2.

* testsuite/gnatprove/tests/T826-013__flow_nested/
Add; doesn't crash, unlike the version with child units.

Change-Id: I9bcf209cf8e405aa8b4397e105da9edc55855469