TA06-012 adapt to From_String returning a valid big number
We should not assume that the return type of the From_String function in
an Integer_Literal or Real_Literal aspect of a type is the type itself.
It is not the case anymore in Big_Reals and Big_Integers as these
functions now return a Valid_Big_Integer/Real.

* src/why/gnat2why-expr.adb
(Transform_Function_Call): Go to the root type to get the hardcoded type.

* tests/T520-027__big_int_literals_bad/test.out
* tests/T520-027__big_real_literals/test.out
Remove always Ttue predicate checks on imprecise calls to From_String
as the return type is now explicitly a valig big number.

Change-Id: Ic46f6cb24c5056b3d3d34618602bb824d440a4c7