T925-002 Cleanup SPARK RM for better readability
Simplify examples so that they are useful for either getting an idea of the
syntax, or for helping in explaining more subtle cases. Avoid very long
examples that are not readable, and best left to the SPARK User's Guide,
or an example elsewhere. Make examples part of the testsuite whenever
possible.

Also cleanup the glossary.

* docs/lrm/source/declarations-and-types.rst
Minor rewording.

* docs/lrm/source/glossary.rst
Rework the incomplete glossary.

* docs/lrm/source/packages.rst
* docs/lrm/source/statements.rst
* docs/lrm/source/subprograms.rst
Simplify examples.

* docs/lrm/source/the-standard-library.rst
Fix the documentation for the support of elementary functions in Numerics.

* testsuite/gnatprove/tests/RM_Examples
Simplify examples used in RM.
Add further examples for the sake of testing.

Change-Id: Ica653712f029cdfd9526802a8ed80a7fe8be06b7