TA06-037 handle hidden parts of abstract states as constituents
In flow we use the abstract state entity to represent either the state
object itself (where none of its constituents are visible) or the hidden
constituents (where only Part_Of constituents are visible). This works,
because most of the flow analysis just deals with "objects".

However, the meaning of an abstract state object matters when we switch
between spec and body visibility, i.e. in the Down_Project/Up_Project
routines, but there we can recover the meaning of those objects with a
visibility query.

This patch fixes the Up_Project routine for dependency contracts, where the
meaning was not taken into account.

* src/flow/flow_refinement.adb
(Up_Project): Handle hidden parts of abstract states as constituents.

* testsuite/gnatprove/tests/TA06-037__depends_child/test.out
Update; spurious check is gone.

Change-Id: I11ff398c6e4a9b9bd9a4ebce3427d941ec7b69bc