T909-033 xfail tests due to gnat sources being out of sync
Disable tests that now produce a SOUNDNESS_BUG marker due to a missing range
check (T909-033__others_2_dim) or are rejected by the frontend due to broken
type resolution (T909-033__2_dim_aggr). Both problems have been already
fixed in the frontend, but the fix are not included in the 21.0 release.

* testsuite/gnatprove/tests/T909-033__2_dim_aggr/test.opt
* testsuite/gnatprove/tests/T909-033__others_2_dim/test.opt
Disable.

Change-Id: Iefdcd3890a09dae13dad6ac2de36d6b8206241f3