Extend "no variable inputs" rules to cover dereferences and generic in-outs.
Object renamings in SPARK require that if the name of the renamed object
is subsequently reevaluated, the result is unchanged. For example, the
legality of a renaming of an element of an array "X : T renames A(I);"
might depend on whether the object I is a constant or a variable.
Extend the rules that implement this design principle to also handle
pointer dereferencing.

A generic actual parameter corresponding to a generic in-out-mode formal
object is similar in many ways to an object renaming declaration (and,
in fact, is implemented by most Ada compilers as an object renaming).
Extend the rules that implement the aforementioned design principle to
also handle this form of "renaming".

* docs/lrms/source/names-and-expressions.rst: Extend the list of
  "no variable inputs allowed" contexts to include a pointer dereference
  prefix in an object renaming declaration; also state that a generic
  actual parameter corresponding to an in-out-mode formal object is
  to be treated like an object renaming with respect to these rules.

* docs/lrm/source/visibility-rules.rst: Correspondingly update the
  non-normative restatement of the "no variable inputs allowed" rules
  for object renamings.

Change-Id: I0b42dd6b138c675912de89aa9e83578e0b07d777
TN: TA07-024