TA01-026 Issue error on unsupported non-static attributes
Attributes which are not supported in GNATprove when their value is
non-static should lead to an error message instead of a crash of the tool.

On the contrary, GNAT implementation-specific attributes which are static
should be supported in GNATprove, they don't cause any portability issue
wrt proof as the compiler also needs to recognize them.

* src/spark/spark_definition.adb
(Mark_Attribute_Reference): Issue error message on unsupported attributes.
(Mark): Remove obsolete and useless marking of original node for literals
 that have been constant folded by the frontend, as it led to spurious
 errors here on static values for supported attributes.

* testsuite/gnatprove/tests/OC06-008__floating_point_attribs/test.opt
* testsuite/gnatprove/tests/OC06-008__floating_point_attribs/test.out
* testsuite/gnatprove/tests/TA01-026__attribute_width/test.opt
* testsuite/gnatprove/tests/TA01-026__attribute_width/test.out
Tests pass now, with error messages.

* gnatprove/tests/MA01-040__flow_attributes_pedantic_warning/test.out
One fewer warning with --pedantic about attribute with implementation-
specific value. This is OK because this specific use of --pedantic is
in no way intended to be complete.

* testsuite/gnatprove/tests/O722-011__standard_attributes/test.out
Test passes now as attributes are static.

* testsuite/gnatprove/tests/P113-023__alias_issue/test.out
Duplicated error message removes as original node is not analyzed anymore.

* testsuite/gnatprove/tests/S729-003__pointer_abstract_state/test.out
Change in error messages as static attribute does not lead to error.

Change-Id: I27102e8f16be99698d8f4a65157110ba6e59a61e