TA20-024 add ghost code to detect the presence of globals in 'Access
As flow analysis does not detect global state in the prefix of the
'Access attribute yet, add ghost code to detect this code in proof.
It should be easier to determine whether or not we are in this case
when encountering a crash.

* src/why/gnat2why-subprograms-pointers.adb
(Transform_Access_Attribute_Of_Subprogram): Add ghost code to check
 whether the prefix of the attribute accesses global data.

Change-Id: I999fe96cf216b61c9967386bc7739d21ada14a0d