TA26-009 mark subtype name in uninitialized allocator
When marking we now examine not just the Retysp of an allocator, but also
the subtype name in an uninitialized allocator. In particular, this is
needed to avoid querying default initialization of a subtype name which is
not in SPARK.

* src/spark/spark_definition.adb
(Mark): Introduce constants for marking uninitialized allocators; check
if the subtype name is in SPARK.

* testsuite/gnatprove/tests/TA26-009__access_init/
Enable and add output; crash is gone.

Change-Id: I8733e9d78c739064f70ce9ce8ef83f2614ed6383