TA28-010 Add test
This is a good test for COLIBRI, which will require improving the its
support as well as the support for Unchecked_Conversion between integers
and floats for provers CVC4, Z3 and COLIBRI.

Intermediate assertions should be provable, as well as AoRTE, but the
postcondition is known to fail, as reported by the user for input value
2**47 + 2**24. This can be checked by executing the postcondition for
this value of V, with a slightly modified code where Correct_Rounding is
defined as a regular function to avoid a GNAT bug that otherwise stops
compilation.

Although the postcondition is not the most precise, it is enough to
detect this failing case. The missing check is for deciding on the correct
rounding in case of tie between two floats, where the one with 0 in the
last place should be chosen. It remains to decide how to best express
this property for provers.

Once we prove all checks here except postcondition, we could attempt
proof of the correct variants (in particular the one that will be selected
for implementation in GNAT).

Change-Id: I4c54483ffd260a89b652e042edab17e108dc9a4c