S820-052 remove unnecessary calls to Unique_Entity
When detecting references to the current type instance within the DIC
expression we don't need Unique_Entity. This is because the pragma Check
expression inside the DIC procedure is referencing the first formal of the
DIC procedure spec, not the first formal of the procedure body.

* src/spark/spark_util-types.adb
(Refers_To_Type_Instance): Remove calls to Unique_Entity.

Change-Id: I598248289eac9e9a6526a679b00013d9e542c68d