TB06-056 add test with wrong folding of operators in loop invariants
Those tests are semantically equivalent, yet the division check is proved in
just one of them. This comes from the loop invariant expression with "/="
being folded into "true" (so it is proved by the "Trivial" prover), while
the expression is ">" is kept as it is.

* testsuite/gnatprove/tests/TB06-056__fold_oper/
Add; both test should be proved.

Change-Id: I2532129099733ef578c72ae484596394edbcd16d