TB12-027 treat abstract subprograms as pure
Abstract subprograms are now considered pure and so are excluded from the GG
machinery.

* src/flow/flow-slice.adb
(Compute_Globals): Ignore calls to abstract subprograms when computing calls
that have to be resolved by the GG machinery for bodies in SPARK.

* src/flow/flow_generated_globals-partial.adb
(Preanalyze_Spec): Likewise for bodies not in SPARK.

* src/flow/flow_utility.adb
(Get_Globals): Explicitly return empty globals for abstract subprograms,
just like we do for pure subprograms.

* testsuite/gnatprove/tests/O504-014__state_machine/
* testsuite/gnatprove/tests/O827-007__dispatch_in_post/
* testsuite/gnatprove/tests/O827-017__dispatching_calls/
Re-enable; crashes are gone, so just accept current outputs.

Change-Id: If3bff9e187c786800c164c3d278544dd2df06232