TA30-014 support iterator filters in flow
This patch implements flow support of iterator filters in quantified
expressions and loop statements. In particular, global variables
referenced in the filter are taken into account in flow analysis.
Detection of variables initialized by the loop is disabled when an
iterator filter is present.

* src/flow/flow-control_flow_graph.adb
(Do_Loop_Statement): Add variables referenced in the filter to the loop
vertex in the CFG. Disable detection of variables initialized by the loop
when a filter is present.

* src/flow/flow_utility.adb
(Get_Variables_Internal): Add the variables referenced in the filter
to the result when called on a quantified expression.

Change-Id: I9c8508fef720146d1ce6095918440ad21acca9b3