TB25-006 stop marking on types derived from types out of SPARK
We don't want to mark types whose Etype is not in SPARK, as the
violations can be duplicates of the violations of the parent. We
could still mark components in tagged extensions, but it seems to
add complexity for little gain.

* src/spark/spark_definition.adb
(Mark_Type_Entity): Stop marking when if the type's Etype is not in
 SPARK.

* tests/O429-044__synchronous_barriers/test.out
Problematic message is not emitted anymore.

* tests/P706-024__private_discr_type/test.out
Message about tagged extension not emitted anymore. Arguably, this is not
an improvement, nor is it really a problem since the type is still rejected
for a good reason and the extension with be rejected too if the Etype is
modified to be in SPARK.

Change-Id: Idfd6661cee4793983ea29d914d28c6f4d9522991