TC10-058 respect informal contract on List_Length
The comment for List_Length says "It is an error to call this function with
No_List (No_List is not considered to be the same as an empty list)." It is
now enforced with an assertion, so this patch adds guards where this
assertion would be failing.

* src/why/gnat2why-expr.adb
(Transform_Record_Aggregate): Don't call List_Length on empty expressions
and associated components of an aggregate.
(Transform_Statement_Or_Declaration): Replace List_Length with a more
efficient Is_Non_Empty_List.

Change-Id: I967ba7d31248ae7400fb7a0341285b96e0269227