TC17-055 in parameter of an access type and frame conditions
When computing the frame condition of a loop, we need to consider that
in parameters of procedures can be modified if the parameter has an
access type.

* src/why/gnat2why-expr-loops-inv.adb
(Update_Status): New parameter Deref_Only to account for a new write
 to the value referenced from an access object only.
(Process_Call_Statement): Call Update_Status on in parameter of an
 access type with Deref_Only set to True.

* tests/TC17-055__crash
Expand and enable test.

Change-Id: I7eb97f3d8010d443d61875ea07e68112c52b8a49