R417-042 check termination of type aspects subprograms
This patch implements the checking of termination of type aspects
(Dynamic_Predicate and Type_Invariant). A check is emitted for each
function referenced in a type aspect that does not terminate.

* docs/lrm/source/packages.rst
Add rule.

* src/flow/flow-analysis-sanity.adb
(Check_Type_Aspects): Check for termination of all functions referenced
in both Dynamic_Predicate and Type_Invariant, if present.

* src/flow/flow_utility.adb
(Collect_Functions_And_Read_Locked_POs): Include functions that are not
visible from the scope in the result. Do not include predicate functions
nor the functions called in their expression in Functions_Called.

* testsuite/gnatprove/tests/P725-016__recursive/
* testsuite/gnatprove/tests/P725-016__recursive_bad_axiom/
* testsuite/gnatprove/tests/Q512-013__flow_type_aspects/
* testsuite/gnatprove/tests/S901-003__flow_term_pred/
* testsuite/gnatprove/tests/T514-020__callback/
Update outputs.

* testsuite/gnatprove/tests/R417-042__flow_type_aspects/
Add test files.

Change-Id: Icdfe2b719690b0a2da67d4b8069aa88ed7daba4f