TA01-035 remove Rep parameter of Apply_Compile_Time_Constraint_Error
In routine Apply_Compile_Time_Constraint_Error the Rep parameter was
disabling rewriting into N_Raise_Constraint_Error node. It is only
referenced when this routine is called from GNATprove, but in GNATprove_Mode
rewriting is disabled anyway.

It seems cleaner to remove this parameter and keep this routine as simple as
possible.

* src/spark/spark_definition.adb
(Mark): Remove parameter Rep in call to Apply_Compile_Time_Constraint_Error
when flagging statically missing components.

Change-Id: I5caba94b943669d7d4156c75cb93b941e27702f1