U111-014 borrow checking of subprograms in protected units
In the borrow checker, the environment for the protected object must be
set not only in protected operations but also on operations nested
directly inside a protected object. On subprograms nested further
inside operations in a protected object, components of the protected
object are handled as global variables.

* src/why/gnat2why-borrow_checker.adb
(Check_Callable_Body): Setup protected components for all operations
 declared directly in a protected type.

Change-Id: I3e258f0b9e0e53c2ffddca6d9c014c228bf9f1b7