TB16-061 rewrite Is_Mutable_In_Why in pattern-matching style
Routine Is_Mutable_In_Why, which uses Is_Constant_In_SPARK, was written as a
sequence of conditions with non-trivial order dependencies. This was causing
some unnecessary computations and generally making this routine hard to
understand and modify.

Now it is rewritten in a pattern-matching style, so handling of individual
object kinds is much clearer. Otherwise its behaviour is not affected.

* src/why/gnat2why-util.adb
(Is_Mutable_In_Why): Rewritten.

Change-Id: Id5cc3b3233a4f3cc6cf9798ef873478722bf5547