U113-026 allow 'address inside address clauses
Previously, 'Address was allowed only at the top level of an address
clause. Handling of address clauses would skip this top-level
'Address attribute, if present, and translate the rest of the
expression (for run-time errors only).

In this patch we remove the "skipping" part and add handling in
Transform_Expr to deal with 'address on any expression. We only care
about runtime errors, so we just evaluate the prefix of the
expression and return a dummy value.

We cannot support the use of 'address properly outside the program
domain, so we check during marking that the context of the 'Address
expression does not require switching domains during proof.

* spark_definition.adb
(Mark_Attribute_Reference): check if 'address appears in supported
  context
* gnat2why-expr.adb:
(Transform_Attr): translate 'address assuming the program domain. we
  only care about RTE.

Change-Id: I81f8e77282dd35ec9105b67d79cd01c60df1dcba